+++ Firmware/LogicAnalyzer_V2/publish.sh
@@ -2,7 +2,8 @@
 
 PKG_NAME="LogicAnalyzer"
 
-BOARD_TYPES="BOARD_PICO BOARD_PICO_W BOARD_PICO_W_WIFI BOARD_ZERO BOARD_PICO_2"
+#BOARD_TYPES="BOARD_PICO BOARD_PICO_W BOARD_PICO_W_WIFI BOARD_ZERO BOARD_PICO_2 
+BOARD_TYPES="BOARD_PICO_ANALYZER BOARD_PICO_ANALYZER2"
 
 PUBLISH_DIR="publish"
 BUILD_DIR="build"
@@ -11,16 +12,17 @@ SOURCE_DIR="."
 BUILD_LOG="$BUILD_DIR/build.log"
 
 # Clean publish directory
-rm --recursive --force $PUBLISH_DIR/*
+rm --recursive --force ${PUBLISH_DIR:?}/*
 mkdir --parents $PUBLISH_DIR
 
 # Create build directory if it does not exists
 mkdir --parents $BUILD_DIR
 
-# Backup old build log
-cp $BUILD_LOG "$BUILD_LOG.bak"
-echo -n "Build Time: " > $BUILD_LOG
-echo $(date) >> $BUILD_LOG
+# Backup old build log if it exists
+[[ -f $BUILD_LOG ]] && cp $BUILD_LOG "$BUILD_LOG.bak"
+
+# Create new build log with start time
+{ echo -n "Build Time: "; date; } > $BUILD_LOG
 
 # Loop through each board type and turbo mode
 for BOARD_TYPE in $BOARD_TYPES; do
@@ -36,7 +38,7 @@ for BOARD_TYPE in $BOARD_TYPES; do
         echo "BOARD_TYPE: $BOARD_TYPE - TURBO_MODE: $TURBO_MODE" >> $BUILD_LOG
 
         # Run CMake config command
-        cmake -GNinja -DBOARD_TYPE=$BOARD_TYPE -DTURBO_MODE=$TURBO_MODE -B $BUILD_DIR -S $SOURCE_DIR &>>$BUILD_LOG
+        cmake -GNinja -DBOARD_TYPE="$BOARD_TYPE" -DTURBO_MODE="$TURBO_MODE" -B $BUILD_DIR -S $SOURCE_DIR &>>$BUILD_LOG
 
         # Run CMake build command
         cmake --build $BUILD_DIR --config Release &>>$BUILD_LOG
